Split and expand tests for selection order viewer
Move focused coverage out of the large GestureSesh test module into selection-order, session-shortcut, and image-loader test files.

Add regression coverage for selection-order helpers, dialog mutation rules, shortcut registration, image-loader cache behavior, and viewport-limited thumbnail loading.
